refactor(threads): isolate semantic event projection
